+++ package.json
@@ -4,7 +4,7 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "dev:seed": "node dbMigration/migrateAndSeed.js",
+    "dev:seed": "node src/dbMigration/migrateAndSeed.js",
     "test": "echo \"Error: no test specified\" && exit 1",
     "start": "nodemon --exec babel-node src/migrateAndSeed.js"
   },
