+++ src/model/associateConfig.js
