+++ src/model/entities/Bookmark.js
