+++ src/model/entities/Image.js
