+++ src/model/entities/Listings.js
